Add likes to the app
