+++ src/modules/display.js
@@ -1,32 +1,5 @@
 import itemCounter from './itemcounter.js';
-
-const likeSend = async (id) => {
-  await fetch(
-    'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/Ntt9tCZkqdoaDg3jriZb/likes', {
-      method: 'POST',
-      body: JSON.stringify({
-        item_id: id,
-      }),
-      headers: {
-        'Content-type': 'application/json; charset=UTF-8',
-      },
-    },
-  );
-};
-const getLike = async () => {
-  const response = await fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/Ntt9tCZkqdoaDg3jriZb/likes', {
-    method: 'GET',
-    headers: {
-      'Content-type': 'application/json; charset=UTF-8',
-    },
-  });
-  const countLike = document.querySelector('#countlike');
-  const data = response.json();
-  countLike.innerHTML = '';
-  data.forEach((el) => {
-    countLike.innerHTML += `${el.item_id} likes`;
-  });
-};
+import likeSend from './sendlikes.js';
 
 const display = (array) => {
   const main = document.querySelector('.main');
@@ -38,10 +11,10 @@ const display = (array) => {
     <div class="comment-like">
       <button class="comment-button ${item.id}" >Comments</button>
       <div class="like-div">
-      <button class="like-button" id='${item.id}'>
+      <button class="like-button ${item.id}" id='${item.id}'>
       <i class="fa fa-heart"></i>
       </button>
-      <p id="countlike">Likes</p>
+      <p id="countlike"></p>
       </div>
     </div>
   </div>`;
@@ -52,12 +25,4 @@ const display = (array) => {
   localStorage.setItem('allGames', JSON.stringify(array));
 };
 
-const likeButtons = document.querySelectorAll('.like-button');
-likeButtons.forEach((button) => {
-  button.addEventListener('click', async (e) => {
-    const { id } = e.target;
-    await likeSend(id);
-    getLike();
-  });
-});
-export default display;
+export { display, likeSend };
