+++ src/modules/sendlikes.js
@@ -1,0 +1,18 @@
+import getLike from './getLikes.js';
+
+const likeSend = async (id) => {
+  await fetch(
+    'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/ngPmi6EPTDMdfkgNT9zd/likes', {
+      method: 'POST',
+      body: JSON.stringify({
+        item_id: id,
+      }),
+      headers: {
+        'Content-type': 'application/json; charset=UTF-8',
+      },
+    },
+  );
+  getLike(id);
+};
+
+export default likeSend;
